split setup_generation file into multiple files
